6542: cleaning code
